+++ src/CommandLineUtils/Attributes/CommandAttribute.cs
@@ -12,6 +12,8 @@ namespace McMaster.Extensions.CommandLineUtils
     [AttributeUsage(AttributeTargets.Class)]
     public sealed class CommandAttribute : Attribute
     {
+        private bool? _clusterOptions;
+
         /// <summary>
         /// Initializes a new <see cref="CommandAttribute"/>.
         /// </summary>
@@ -78,6 +80,23 @@ namespace McMaster.Extensions.CommandLineUtils
         /// </summary>
         public CultureInfo ParseCulture { get; set; } = CultureInfo.CurrentCulture;
 
+        /// <summary>
+        /// <para>
+        /// One or more options of <see cref="CommandOptionType.NoValue"/>, followed by at most one option that takes values, should be accepted when grouped behind one '-' delimiter.
+        /// </para>
+        /// <para>
+        /// This defaults to true unless an option with a short name of two or more characters is added.
+        /// </para>
+        /// <para>
+        /// See <see cref="ParserSettings.ClusterOptions"/> for more details.
+        /// </para>
+        /// </summary>
+        public bool ClusterOptions
+        {
+            get => _clusterOptions ?? true;
+            set => _clusterOptions = value;
+        }
+
         internal void Configure(CommandLineApplication app)
         {
             // this might have been set from SubcommandAttribute
@@ -92,6 +111,12 @@ namespace McMaster.Extensions.CommandLineUtils
             app.ThrowOnUnexpectedArgument = ThrowOnUnexpectedArgument;
             app.OptionsComparison = OptionsComparison;
             app.ValueParsers.ParseCulture = ParseCulture;
+
+            // detect if this is explicitly set. Otherwise, we try to infer if clustering options is okay.
+            if (_clusterOptions.HasValue)
+            {
+                app.ParserSettings.ClusterOptions = ClusterOptions;
+            }
         }
     }
 }