+++ src/CommandLineUtils/ParserSettings.cs
@@ -1,0 +1,47 @@
+﻿// Copyright (c) Nate McMaster.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace McMaster.Extensions.CommandLineUtils
+{
+    /// <summary>
+    /// Settings which control the parser behavior.
+    /// </summary>
+    public class ParserSettings
+    {
+        /// <summary>
+        /// <para>
+        /// One or more options of <see cref="CommandOptionType.NoValue"/>, followed by at most one option that takes values, should be accepted when grouped behind one '-' delimiter.
+        /// </para>
+        /// <para>
+        /// When true, the following are equivalent.
+        ///
+        /// <code>
+        /// -abcXyellow
+        /// -abcX=yellow
+        /// -abcX:yellow
+        /// -abc -X=yellow
+        /// -ab -cX=yellow
+        /// -a -b -c -Xyellow
+        /// -a -b -c -X yellow
+        /// -a -b -c -X=yellow
+        /// -a -b -c -X:yellow
+        /// </code>
+        /// </para>
+        /// <para>
+        /// This defaults to true unless an option with a short name of two or more characters is added.
+        /// </para>
+        /// </summary>
+        /// <remarks>
+        /// <seealso href="https://www.gnu.org/software/libc/manual/html_node/Argument-Syntax.html"/>
+        /// </remarks>
+        public bool ClusterOptions
+        {
+            get => _clusterOptions ?? true;
+            set => _clusterOptions = value;
+        }
+
+        private bool? _clusterOptions;
+
+        internal bool ClusterOptionsWasSetExplicitly => _clusterOptions.HasValue;
+    }
+}