+++ test/CommandLineUtils.Tests/ResponseFileTests.cs
@@ -189,8 +189,10 @@ namespace McMaster.Extensions.CommandLineUtils.Tests
         [InlineData("a \"\" b", new[] { "a", "", "b" })]
         [InlineData("\"double quoted arg with 'single quotes'\"", new[] { "double quoted arg with 'single quotes'" })]
         [InlineData("'single quoted arg with \"double quotes\"'", new[] { "single quoted arg with \"double quotes\"" })]
-        [InlineData("\"double quoted arg with escaped \\\" double quote\"", new[] { "double quoted arg with escaped \" double quote" })]
-        [InlineData("'single quoted arg with escaped \\' single quote'", new[] { "single quoted arg with escaped ' single quote" })]
+        [InlineData("\"double quoted arg with escaped \\\" double quote\"",
+            new[] { "double quoted arg with escaped \" double quote" })]
+        [InlineData("'single quoted arg with escaped \\' single quote'",
+            new[] { "single quoted arg with escaped ' single quote" })]
         [InlineData(@"\", new[] { @"\" })]
         [InlineData(@"\\", new[] { @"\\" })]
         [InlineData("\\\"", new[] { "\"" })]
@@ -241,6 +243,21 @@ namespace McMaster.Extensions.CommandLineUtils.Tests
         }
 
         [Fact]
+        public void ItHandlesArgumentSeparatorInResponseFile()
+        {
+            var app = new CommandLineApplication
+            {
+                ThrowOnUnexpectedArgument = false,
+                ResponseFileHandling = ResponseFileHandling.ParseArgsAsSpaceSeparated,
+                AllowArgumentSeparator = true,
+            };
+            var rspFile = CreateResponseFile("-- --hello");
+            app.Execute("@" + rspFile, "--world");
+
+            Assert.Equal(new[] { "--hello", "--world" }, app.RemainingArguments);
+        }
+
+        [Fact]
         public void SubcommandsCanResponseFileOptions()
         {
             var app = new CommandLineApplication();
@@ -253,10 +270,10 @@ namespace McMaster.Extensions.CommandLineUtils.Tests
             var rspFile = CreateResponseFile(" 'lorem ipsum' ", "dolor sit amet");
             app.Execute("save", "@" + rspFile);
             Assert.Collection(wordArgs.Values,
-               a => Assert.Equal("lorem ipsum", a),
-               a => Assert.Equal("dolor", a),
-               a => Assert.Equal("sit", a),
-               a => Assert.Equal("amet", a));
+                a => Assert.Equal("lorem ipsum", a),
+                a => Assert.Equal("dolor", a),
+                a => Assert.Equal("sit", a),
+                a => Assert.Equal("amet", a));
         }
 
         [Fact]
@@ -273,9 +290,9 @@ namespace McMaster.Extensions.CommandLineUtils.Tests
             Assert.Equal("lorem ipsum", opt.Value());
 
             Assert.Collection(app.RemainingArguments,
-               a => Assert.Equal("dolor", a),
-               a => Assert.Equal("sit", a),
-               a => Assert.Equal("amet", a));
+                a => Assert.Equal("dolor", a),
+                a => Assert.Equal("sit", a),
+                a => Assert.Equal("amet", a));
         }
     }
 }